# openapi.yaml
openapi: 3.1.0
info:
  title: Notifier
  description: Internal alerting service for Watchdog
  version: 0.0.3
paths:
...
  /internal/v1/api/alertmanager/incidents:
    get:
      tags:
      - alertmanager-incidents
      summary: List Incidents
      description: Lists alert incidents visible to the current user with optional
        status, visibility, and group filters.
      operationId: list_incidents
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: visibility
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Visibility
      - name: group_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: The incidents visible to the current caller.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertIncident'
                title: Response List Incidents
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /internal/v1/api/alertmanager/incidents/summary:
    get:
      tags:
      - alertmanager-incidents
      summary: Get Incident Summary
      description: Returns an aggregated summary of incidents visible to the current
        user.
      operationId: get_incident_summary
      responses:
        '200':
          description: Aggregated incident summary counts and breakdowns.
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Get Incident Summary
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
      security:
      - HTTPBearer: []
  /internal/v1/api/alertmanager/incidents/{incident_id}:
    patch:
      tags:
      - alertmanager-incidents
      summary: Update Incident
      description: Updates an incident's status, assignment, Jira metadata, or visibility
        settings.
      operationId: update_incident
      security:
      - HTTPBearer: []
      parameters:
      - name: incident_id
        in: path
        required: true
        schema:
          type: string
          title: Incident Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertIncidentUpdateRequest'
      responses:
        '200':
          description: The updated incident.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertIncident'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '429':
          description: Too Many Requests
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AlertIncident:
      properties:
        id:
          type: string
          title: Id
          examples:
          - incident-123
        fingerprint:
          type: string
          title: Fingerprint
          examples:
          - 01ARZ3NDEKTSV4RRFFQ69G5FAV
        alertName:
          type: string
          title: Alertname
          examples:
          - HighCpuUsage
        severity:
          type: string
          title: Severity
          examples:
          - critical
        status:
          $ref: '#/components/schemas/IncidentStatus'
          examples:
          - open
        assignee:
          anyOf:
          - type: string
          - type: 'null'
          title: Assignee
          examples:
          - alice@example.com
        notes:
          items:
            $ref: '#/components/schemas/IncidentNote'
          type: array
          title: Notes
        labels:
          additionalProperties:
            type: string
          type: object
          title: Labels
          examples:
          - service: api
            severity: critical
        annotations:
          additionalProperties:
            type: string
          type: object
          title: Annotations
          examples:
          - summary: CPU above 95%
        visibility:
          $ref: '#/components/schemas/IncidentVisibility'
          default: public
          examples:
          - public
        sharedGroupIds:
          items:
            type: string
          type: array
          title: Sharedgroupids
          examples:
          - - group-ops
        jiraTicketKey:
          anyOf:
          - type: string
          - type: 'null'
          title: Jiraticketkey
          examples:
          - OPS-321
        jiraTicketUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Jiraticketurl
          examples:
          - https://jira.example.internal/browse/OPS-321
        jiraIntegrationId:
          anyOf:
          - type: string
          - type: 'null'
          title: Jiraintegrationid
          examples:
          - jira-int-01
        startsAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Startsat
          examples:
          - '2026-04-03T11:55:00Z'
        lastSeenAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastseenat
          examples:
          - '2026-04-03T12:05:00Z'
        resolvedAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolvedat
          examples:
          - '2026-04-03T12:20:00Z'
        createdAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdat
          examples:
          - '2026-04-03T12:00:00Z'
        updatedAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Updatedat
          examples:
          - '2026-04-03T12:10:00Z'
        userManaged:
          type: boolean
          title: Usermanaged
          default: false
          examples:
          - false
        hideWhenResolved:
          type: boolean
          title: Hidewhenresolved
          default: false
          examples:
          - false
      type: object
      required:
      - id
      - fingerprint
      - alertName
      - severity
      - status
      - lastSeenAt
      - createdAt
      - updatedAt
      title: AlertIncident
    AlertIncidentUpdateRequest:
      properties:
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
          examples:
          - resolved
        assignee:
          anyOf:
          - type: string
          - type: 'null'
          title: Assignee
          examples:
          - alice@example.com
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
          examples:
          - Resolved after scaling the deployment
        actorUsername:
          anyOf:
          - type: string
          - type: 'null'
          title: Actorusername
          examples:
          - alice
        visibility:
          anyOf:
          - $ref: '#/components/schemas/IncidentVisibility'
          - type: 'null'
          examples:
          - group
        sharedGroupIds:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Sharedgroupids
          examples:
          - - group-ops
        jiraTicketKey:
          anyOf:
          - type: string
          - type: 'null'
          title: Jiraticketkey
          examples:
          - OPS-321
        jiraTicketUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Jiraticketurl
          examples:
          - https://jira.example.internal/browse/OPS-321
        jiraIntegrationId:
          anyOf:
          - type: string
          - type: 'null'
          title: Jiraintegrationid
          examples:
          - jira-int-01
        hideWhenResolved:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Hidewhenresolved
          examples:
          - true
      type: object
      title: AlertIncidentUpdateRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IncidentNote:
      properties:
        author:
          type: string
          title: Author
          examples:
          - alice@example.com
        text:
          type: string
          title: Text
          examples:
          - Investigating elevated CPU on api-01
        createdAt:
          type: string
          title: Createdat
          examples:
          - '2026-04-03T12:00:00Z'
      type: object
      required:
      - author
      - text
      - createdAt
      title: IncidentNote
    IncidentStatus:
      type: string
      enum:
      - open
      - resolved
      title: IncidentStatus
    IncidentVisibility:
      type: string
      enum:
      - public
      - private
      - group
      title: IncidentVisibility
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...